added flags

+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: gadeneux <gadeneux@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/16 11:59:27 by gadeneux          #+#    #+#             */
-/*   Updated: 2022/05/30 18:14:26 by gadeneux         ###   ########.fr       */
+/*   Updated: 2022/06/08 15:19:44 by gadeneux         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "header.h"
 
-size_t	ft_strlen(char *s)
+int	ft_strlen(char *s)
 {
 	int	i;
