+++ header.h
@@ -6,7 +6,7 @@
 /*   By: gadeneux <gadeneux@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/05 16:23:29 by gadeneux          #+#    #+#             */
-/*   Updated: 2022/06/07 22:33:24 by gadeneux         ###   ########.fr       */
+/*   Updated: 2022/06/08 15:20:07 by gadeneux         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -115,7 +115,7 @@ char	*ft_strdup2(char *s1, int len);
 char	*ft_strjoin(char *s1, char *s2);
 void	*ft_memcpy(void *dst, void *src, size_t n);
 size_t	ft_strlcat(char *dst, char *src, size_t dstsize);
-size_t	ft_strlen(char *s);
+int		ft_strlen(char *s);
 int	ft_str_contains(char *str, char c);
 
 int		ft_str_isempty(char *str, int begin, int len);
